Fix case on objects and classes
